# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Verisafe service API",
        "title": "Verisafe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Open Crafts Interactive",
            "url": "https://opencrafts.io/about",
            "email": "developers@opencrafts.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "qaverisafe.opencrafts.io",
    "basePath": "/",
    "paths": {
...
        "/oauth/grants": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Pre-flight for the broker: which providers an account has connected and what each grant covers. Never includes tokens.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "List an account's third-party connections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grants.GrantView"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing or malformed account_id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "403": {
                        "description": "Caller is not a service token or lacks the permission",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
...
        "/oauth/{provider}/reconcile": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Refreshes an account's grant purely to learn the scopes the provider reports, converting a presumed scope list into a verified one. Operational endpoint; the background reconciler does this automatically.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "Force a grant to be re-verified against the provider",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "spotify"
                        ],
                        "type": "string",
                        "description": "OAuth2 provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account to reconcile (capabilities ignored)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.ProviderTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Malformed body or bad account_id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "403": {
                        "description": "Lacks manage:provider_token:any",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "404": {
                        "description": "No grant for this account and provider",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "503": {
                        "description": "Provider is temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/oauth/{provider}/token": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Returns a usable provider access token for the given account, refreshing it against the provider if the stored one is stale. Requires a service token (X-API-Key) and the read:provider_token:any permission — a human Bearer JWT is rejected even if it holds the permission. Refresh tokens are never returned. When the user has not granted the requested capabilities, responds 403 with a machine-readable insufficient_scope body naming what is missing and where to send the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "Obtain a third-party OAuth access token on behalf of a user",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "spotify"
                        ],
                        "type": "string",
                        "description": "OAuth2 provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account and capabilities required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.ProviderTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oauth.ProviderTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Malformed body, bad account_id, or unknown capability",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "403": {
                        "description": "Caller is not a service token, lacks the permission, or the user has not granted the capability",
                        "schema": {
                            "$ref": "#/definitions/oauth.InsufficientScopeResponse"
                        }
                    },
                    "404": {
                        "description": "Unknown provider, or the account has never connected it",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "409": {
                        "description": "Provider cannot refresh tokens and the stored one has expired",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "503": {
                        "description": "Provider is temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the provided input is invalid or malformed"
                }
            }
        },
...
        "grants.GrantView": {
            "type": "object",
            "properties": {
                "available_capabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/providers.Capability"
                    }
                },
                "connected_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "external_user_id": {
                    "type": "string"
                },
                "granted_capabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/providers.Capability"
                    }
                },
                "granted_scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "last_refreshed_at": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "refresh_available": {
                    "type": "boolean"
                },
                "revoked": {
                    "type": "boolean"
                },
                "revoked_reason": {
                    "type": "string"
                },
                "scopes_verified": {
                    "type": "boolean"
                },
                "supports_incremental": {
                    "type": "boolean"
                }
            }
        },
        "oauth.InsufficientScopeResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "authorization_body": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "authorization_method": {
                    "type": "string"
                },
                "authorization_url": {
                    "description": "AuthorizationURL points at Verisafe's own scope-upgrade endpoint, not at\nthe provider. It cannot be a directly openable provider URL: starting an\nauthorization requires the *user's* JWT, which a service token holder\ndoes not have. The service relays this to its own client, which calls\nthe endpoint with the user's token and opens the URL it returns.",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "granted_capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "missing_capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "missing_scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "oauth.ProviderTokenRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "capabilities": {
                    "description": "Capabilities are abstract names (\"calendar\"), not raw provider scopes.\nThere is deliberately no raw-scope escape hatch: a caller needing\nsomething unmapped adds it to the provider registry, which keeps the\nregistry authoritative and the insufficient_scope response meaningful.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "oauth.ProviderTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "account_id": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "from_cache": {
                    "type": "boolean"
                },
                "granted_scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "provider": {
                    "type": "string"
                },
                "refreshed": {
                    "type": "boolean"
                },
                "scopes_verified": {
                    "description": "ScopesVerified is false when the scope list is still what Verisafe\npresumed from historical logins rather than what the provider confirmed.\nA caller that gets a 403 from the provider despite this being false\nshould simply re-call the broker: the failed attempt will have converted\nthe grant to verified.",
                    "type": "boolean"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "providers.Capability": {
            "type": "string",
            "enum": [
                "identity",
                "calendar",
                "tasks",
                "playback",
                "playlist",
                "library"
            ],
            "x-enum-varnames": [
                "CapabilityIdentity",
                "CapabilityCalendar",
                "CapabilityTasks",
                "CapabilityPlayback",
                "CapabilityPlaylist",
                "CapabilityLibrary"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}